# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Juridica",
        "version": "1.0 (2024-07-22)"
    },
    "servers": [
        {
            "url": "http://localhost:8081/legal/public/"
        }
    ],
    "paths": {
...
        "/api/person": {
            "get": {
                "tags": [
                    "Person"
                ],
                "summary": "Get all person or Search full_name and Doc",
                "operationId": "74d719561741d8f0cf67db2cd1664ca1",
                "parameters": [
                    {
                        "name": "x-token",
                        "in": "header",
                        "description": "Key API",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "names or doc",
                        "required": false,
                        "example": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                }
            },
            "post": {
                "tags": [
                    "Person"
                ],
                "summary": "Create Person",
                "operationId": "e3fa321f1f16f618321269761b9dce45",
                "parameters": [
                    {
                        "name": "x-token",
                        "in": "header",
                        "description": "Key API",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "examples": {
                                "result": {
                                    "summary": "An result object.",
                                    "value": {
                                        "full_name": "",
                                        "surnames": "",
                                        "type_doc": "",
                                        "number_doc": ""
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                }
            }
        },
        "/api/person/{id}": {
            "put": {
                "tags": [
                    "Person"
                ],
                "summary": "Update Person",
                "operationId": "1a7473b83a56b2325d2df07319dd3140",
                "parameters": [
                    {
                        "name": "x-token",
                        "in": "header",
                        "description": "Key API",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "surnames",
                                    "type_doc",
                                    "number_doc"
                                ],
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "USUARIO"
                                    },
                                    "surnames": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "USUARIO"
                                    },
                                    "type_doc": {
                                        "type": "string",
                                        "maxLength": 5,
                                        "example": "CC"
                                    },
                                    "number_doc": {
                                        "type": "string",
                                        "maxLength": 20,
                                        "example": "1111111111"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recurso actualizado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "Process"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso no encontrado"
                    },
                    "400": {
                        "description": "Solicitud inválida"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Person"
                ],
                "summary": "Delete Process",
                "operationId": "05193f8fab9847a46b83fc7584ac35f0",
                "parameters": [
                    {
                        "name": "x-token",
                        "in": "header",
                        "description": "Key API",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recurso actualizado con éxito"
                    },
                    "404": {
                        "description": "Recurso no encontrado"
                    },
                    "400": {
                        "description": "Solicitud inválida"
                    }
                }
            }
        },
...
    },
...
}